% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fastdmvnorm}
\alias{fastdmvnorm}
\title{Faster multivariate normal density using multiple processor cores}
\usage{
fastdmvnorm(x, mean = rep(0, ncol(x)), sigma)
}
\arguments{
\item{x}{matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{mean}{mean vector, default is rep(0, length = ncol(x)).}

\item{sigma}{covariance matrix}
}
\description{
Faster multivariate normal density using multiple processor cores
}
